+++ plugin/customcommandsplugin/customcommand.go
@@ -1,0 +1,34 @@
+package customcommandsplugin
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/torlenor/abylebotter/model"
+)
+
+func (p *CustomCommandsPlugin) onCustomCommand(post model.Post) {
+	// TODO: Do not fetch them every time
+	customCommands, err := p.getCommands()
+	if err != nil {
+		p.API.LogError(fmt.Sprintf("Could not get custom commands from storage: %s", err))
+		return
+	}
+
+	if len(post.Content) < 2 {
+		p.API.LogWarn("'%s' cannot be a custom command: String too short")
+		return
+	}
+
+	command := post.Content[1:]
+
+	for _, c := range customCommands.Commands {
+		if c.ChannelID != post.ChannelID {
+			continue
+		}
+
+		if strings.HasPrefix(command, c.Command) {
+			p.returnMessage(post.ChannelID, c.Text)
+		}
+	}
+}