+++ storage/mongostorage/store.go
@@ -43,3 +43,15 @@ func (b *MongoStorage) StoreTimedMessagesPluginMessages(botID, pluginID, identif
 
 	return nil
 }
+
+// StoreCustomCommandsPluginCommands stores data for CustomCommandsPlugin
+func (b *MongoStorage) StoreCustomCommandsPluginCommands(botID, pluginID, identifier string, data storagemodels.CustomCommandsPluginCommands) error {
+	c := b.db.Collection(collectionPluginStorage)
+	filter := bson.M{fieldBotID: botID, fieldPluginID: pluginID, fieldIdentifier: identifier}
+	_, err := c.ReplaceOne(context.Background(), filter, customCommandsPluginCommandsData{BotID: botID, PluginID: pluginID, Identifier: identifier, Data: data}, options.Replace().SetUpsert(true))
+	if err != nil {
+		return err
+	}
+
+	return nil
+}